Don't overwrite symbol information if the method has been modified.
Reading symbols will try to fill in symbols for things which don't have them. Recent change made it so that symbols are read even for methods which do have debug info, but no sequence points.

For tools which edit method bodies, lack of sequence points might be a real scenario (if the method body is completely overwritten, the tool might not have any valid information to put into the sequence points).

So avoid overwriting the debug info in case this happened.